# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Office management control system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Staff" }
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateStaffBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Staff" }
  /staff/{id}:
    get:
      operationId: getStaff
      tags: [staff]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Staff
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Staff" }
    patch:
      operationId: updateStaff
      tags: [staff]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateStaffBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Staff" }
    delete:
      operationId: deleteStaff
      tags: [staff]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ---------- Clients ----------
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      parameters:
        - in: query
          name: assigneeId
          schema: { type: integer }
        - in: query
          name: clientId
          schema: { type: integer }
        - in: query
          name: status
          schema:
            type: string
            enum: [todo, in_progress, in_review, done]
        - in: query
          name: priority
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
    post:
      operationId: createTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTaskBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTaskBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ---------- Attendance ----------
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      parameters:
        - in: query
          name: unreadOnly
          schema: { type: boolean }
      responses:
        "200":
          description: Notifications for current user
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Notification" }
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }
...
components:
  schemas:
...
    Staff:
      type: object
      required: [id, email, name, role, status, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role:
          type: string
          enum: [admin, staff]
        status:
          type: string
          enum: [active, inactive]
        position: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        department: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
        joinedAt: { type: string, nullable: true, format: date }
        createdAt: { type: string, format: date-time }
    CreateStaffBody:
      type: object
      required: [email, name, role, password]
      properties:
        email: { type: string, format: email }
        name: { type: string, minLength: 1 }
        role:
          type: string
          enum: [admin, staff]
        password: { type: string, minLength: 6 }
        position: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        department: { type: string, nullable: true }
        joinedAt: { type: string, nullable: true, format: date }
    UpdateStaffBody:
      type: object
      properties:
        name: { type: string }
        role:
          type: string
          enum: [admin, staff]
        status:
          type: string
          enum: [active, inactive]
        position: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        department: { type: string, nullable: true }
        password: { type: string, minLength: 6, nullable: true }
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string, nullable: true }
        status:
          type: string
          enum: [todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId: { type: integer, nullable: true }
        assigneeName: { type: string, nullable: true }
        clientId: { type: integer, nullable: true }
        clientName: { type: string, nullable: true }
        dueDate: { type: string, nullable: true, format: date }
        createdById: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, nullable: true, format: date-time }
    CreateTaskBody:
      type: object
      required: [title, status, priority]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string, nullable: true }
        status:
          type: string
          enum: [todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId: { type: integer, nullable: true }
        clientId: { type: integer, nullable: true }
        dueDate: { type: string, nullable: true, format: date }
    UpdateTaskBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        status:
          type: string
          enum: [todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId: { type: integer, nullable: true }
        clientId: { type: integer, nullable: true }
        dueDate: { type: string, nullable: true, format: date }
...
    Notification:
      type: object
      required: [id, userId, title, message, type, read, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        title: { type: string }
        message: { type: string }
        type:
          type: string
          enum: [task, invoice, attendance, system, reminder]
        read: { type: boolean }
        link: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...